[BACKLOG-2905]Set default TTL for Service Cache
