+++ src/main/java/org/pentaho/di/trans/dataservice/optimization/cache/ui/ServiceCacheController.java
@@ -24,23 +24,30 @@ package org.pentaho.di.trans.dataservice.optimization.cache.ui;
 
 import org.pentaho.di.trans.dataservice.optimization.PushDownOptimizationMeta;
 import org.pentaho.di.trans.dataservice.optimization.cache.ServiceCache;
-import org.pentaho.di.trans.dataservice.ui.DataServiceDialog;
+import org.pentaho.di.trans.dataservice.optimization.cache.ServiceCacheFactory;
 import org.pentaho.di.trans.dataservice.ui.controller.AbstractController;
+import org.pentaho.di.trans.dataservice.ui.model.DataServiceModel;
 import org.pentaho.ui.xul.binding.Binding;
 import org.pentaho.ui.xul.binding.BindingFactory;
 import org.pentaho.ui.xul.components.XulCheckbox;
 import org.pentaho.ui.xul.components.XulTextbox;
+
+import java.util.List;
 
 /**
  * @author nhudak
  */
 public class ServiceCacheController extends AbstractController {
   private static final String NAME = "serviceCacheCtrl";
-  private final DataServiceDialog dialog;
+  private final ServiceCacheFactory factory;
 
-  public ServiceCacheController( DataServiceDialog dialog ) {
-    this.dialog = dialog;
+  public ServiceCacheController( ServiceCacheFactory factory ) {
+    this.factory = factory;
     setName( NAME );
+  }
+
+  public void initBindings( DataServiceModel model ) {
+    initBindings( locateServiceCacheMeta( model ) );
   }
 
   public void initBindings( PushDownOptimizationMeta meta ) {
@@ -55,7 +62,38 @@ public class ServiceCacheController extends AbstractController {
     checkbox.setChecked( meta.isEnabled() );
     bindingFactory.createBinding( checkbox, "checked", meta, "enabled" );
 
-    ttl.setValue( serviceCache.getTimeToLive() );
+    try {
+      ttl.setValue( serviceCache.getConfiguredTimeToLive() );
+    } catch ( Exception e ) {
+      getLogChannel().logError( "Unable to set default TTL", e );
+    }
     bindingFactory.createBinding( ttl, "value", serviceCache, "timeToLive" );
+  }
+
+  /**
+   * Locate or create a pushdown optimization for service cache. Only one should exist, others will be removed if found.
+   *
+   * @param model Data Service model to update
+   * @return The ONLY Optimization Meta with a Service Cache type
+   */
+  protected PushDownOptimizationMeta locateServiceCacheMeta( DataServiceModel model ) {
+    List<PushDownOptimizationMeta> cacheOptimizations = model.getPushDownOptimizations( factory.getType() );
+
+    PushDownOptimizationMeta meta;
+    if ( cacheOptimizations.isEmpty() ) {
+      meta = new PushDownOptimizationMeta();
+      meta.setStepName( model.getServiceStep() );
+      meta.setType( factory.createPushDown() );
+
+      model.add( meta );
+    } else {
+      meta = cacheOptimizations.get( 0 );
+    }
+
+    if ( cacheOptimizations.size() > 1 ) {
+      model.removeAll( cacheOptimizations.subList( 1, cacheOptimizations.size() ) );
+    }
+
+    return meta;
   }
 }
