+++ src/main/java/org/pentaho/di/trans/dataservice/optimization/cache/ui/ServiceCacheOverlay.java
@@ -23,12 +23,8 @@
 package org.pentaho.di.trans.dataservice.optimization.cache.ui;
 
 import org.pentaho.di.core.exception.KettleException;
-import org.pentaho.di.trans.dataservice.optimization.PushDownOptimizationMeta;
 import org.pentaho.di.trans.dataservice.optimization.cache.ServiceCacheFactory;
 import org.pentaho.di.trans.dataservice.ui.DataServiceDialog;
-import org.pentaho.di.trans.dataservice.ui.model.DataServiceModel;
-
-import java.util.List;
 
 /**
  * @author nhudak
@@ -48,37 +44,11 @@ public class ServiceCacheOverlay implements DataServiceDialog.OptimizationOverla
   }
 
   @Override public void apply( DataServiceDialog dialog ) throws KettleException {
-    ServiceCacheController controller = new ServiceCacheController( dialog );
+    ServiceCacheController controller = new ServiceCacheController( factory );
 
     dialog.applyOverlay( getClass().getClassLoader(), XUL_OVERLAY ).addEventHandler( controller );
 
-    controller.initBindings( locateServiceCacheMeta( dialog.getModel() ) );
+    controller.initBindings( dialog.getModel() );
   }
 
-  /**
-   * Locate or create a pushdown optimization for service cache. Only one should exist, others will be removed if found.
-   *
-   * @param model Data Service model to update
-   * @return The ONLY Optimization Meta with a Service Cache type
-   */
-  protected PushDownOptimizationMeta locateServiceCacheMeta( DataServiceModel model ) {
-    List<PushDownOptimizationMeta> cacheOptimizations = model.getPushDownOptimizations( factory.getType() );
-
-    PushDownOptimizationMeta meta;
-    if ( cacheOptimizations.isEmpty() ) {
-      meta = new PushDownOptimizationMeta();
-      meta.setStepName( model.getServiceStep() );
-      meta.setType( factory.createPushDown() );
-
-      model.add( meta );
-    } else {
-      meta = cacheOptimizations.get( 0 );
-    }
-
-    if ( cacheOptimizations.size() > 1 ) {
-      model.removeAll( cacheOptimizations.subList( 1, cacheOptimizations.size() ) );
-    }
-
-    return meta;
-  }
 }
